+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -348,6 +348,7 @@ class ConfigurationTest extends TestCase
                 'transports' => [],
                 'enabled' => !class_exists(FullStack::class) && class_exists(Mailer::class),
             ],
+            'error_controller' => 'error_controller',
         ];
     }
 }