+++ src/Symfony/Component/Validator/Mapping/Cache/DoctrineCache.php
@@ -14,10 +14,14 @@ namespace Symfony\Component\Validator\Mapping\Cache;
 use Doctrine\Common\Cache\Cache;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 
+@trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.4.', DoctrineCache::class), E_USER_DEPRECATED);
+
 /**
  * Adapts a Doctrine cache to a CacheInterface.
  *
  * @author Florian Voutzinos <florian@voutzinos.com>
+ *
+ * @deprecated since Symfony 4.4.
  */
 final class DoctrineCache implements CacheInterface
 {