+++ src/Symfony/Component/Validator/Tests/Mapping/Cache/DoctrineCacheTest.php
@@ -14,6 +14,9 @@ namespace Symfony\Component\Validator\Tests\Mapping\Cache;
 use Doctrine\Common\Cache\ArrayCache;
 use Symfony\Component\Validator\Mapping\Cache\DoctrineCache;
 
+/**
+ * @group legacy
+ */
 class DoctrineCacheTest extends AbstractCacheTest
 {
     protected function setUp(): void