+++ src/Symfony/Component/Validator/Tests/ValidatorBuilderTest.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\Validator\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Component\Validator\ValidatorBuilder;
 
 class ValidatorBuilderTest extends TestCase
@@ -83,11 +84,9 @@ class ValidatorBuilderTest extends TestCase
         $this->assertSame($this->builder, $this->builder->disableAnnotationMapping());
     }
 
-    public function testSetMetadataCache()
+    public function testSetMappingCache()
     {
-        $this->assertSame($this->builder, $this->builder->setMetadataCache(
-            $this->getMockBuilder('Symfony\Component\Validator\Mapping\Cache\CacheInterface')->getMock())
-        );
+        $this->assertSame($this->builder, $this->builder->setMappingCache($this->createMock(CacheItemPoolInterface::class)));
     }
 
     public function testSetConstraintValidatorFactory()